Hoist CollectionScreen and ItemDetailScreen out of AppContent (CUR-149)
Both screens were declared inline inside AppContent, so every app-level
state change (per-keystroke save status, sync flips, toast dismissals)
created new component identities and remounted the active screen. Typing
into Item Detail committed one character and dropped focus; collection
search was wiped whenever a background sync settled.

The screens now live in components/ as top-level components (matching
HomeScreen) with their AppContent closures converted to props. The
retry-save and delete-collection flows stay in AppContent as callbacks so
the save-revision refs and pending-save bookkeeping remain private to it.

Extraction also surfaced a latent rules-of-hooks bug: two ItemDetailScreen
effects ran after the loading early-return, which only worked because the
screen remounted every render. They now run before it.

Adds character-by-character typing regression tests (all four fail
against the previous inline declarations).
